requirements.txt
Flask==3.1.2
fpdf2==2.8.5
Pillow==11.3.0
qrcode==8.2
